+++ src/main/kotlin/com/github/jensim/megamanipulator/actions/search/sourcegraph/SourcegraphSearchClient.kt
@@ -1,64 +1,74 @@
 package com.github.jensim.megamanipulator.actions.search.sourcegraph
 
+import com.expediagroup.graphql.client.ktor.GraphQLKtorClient
+import com.expediagroup.graphql.client.serializer.GraphQLClientSerializer
+import com.expediagroup.graphql.client.types.GraphQLClientRequest
+import com.expediagroup.graphql.client.types.GraphQLClientResponse
 import com.github.jensim.megamanipulator.actions.NotificationsOperator
 import com.github.jensim.megamanipulator.actions.search.SearchResult
+import com.github.jensim.megamanipulator.graphql.generated.sourcegraph.Search
+import com.github.jensim.megamanipulator.graphql.generated.sourcegraph.Search.Result
+import com.github.jensim.megamanipulator.graphql.generated.sourcegraph.search.CommitSearchResult
+import com.github.jensim.megamanipulator.graphql.generated.sourcegraph.search.FileMatch
+import com.github.jensim.megamanipulator.graphql.generated.sourcegraph.search.Repository2
 import com.github.jensim.megamanipulator.http.HttpClientProvider
 import com.github.jensim.megamanipulator.settings.SearchHostSettings.SourceGraphSettings
 import com.intellij.notification.NotificationType.ERROR
 import com.intellij.notification.NotificationType.WARNING
 import com.jetbrains.rd.util.printlnError
-import io.ktor.client.features.timeout
-import io.ktor.client.request.post
-import io.ktor.client.statement.HttpResponse
-import io.ktor.client.statement.readText
-import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
+import org.slf4j.LoggerFactory
+import java.net.URL
+import kotlin.reflect.KClass
 
 class SourcegraphSearchClient(
     private val httpClientProvider: HttpClientProvider,
     private val notificationsOperator: NotificationsOperator,
-    private val json: Json,
 ) {
+
+    private val log = LoggerFactory.getLogger(this.javaClass)
 
     suspend fun search(searchHostName: String, settings: SourceGraphSettings, search: String): Set<SearchResult> {
         try {
             val raw = rawSearch(searchHostName, settings, search)
-            if ("${raw.status.value}".take(1) != "2") {
-                throw RuntimeException("${raw.status} ${raw.readText()}")
+            if (!raw.errors.isNullOrEmpty()) {
+                throw RuntimeException("${raw.errors}")
             }
-            val response: SourcegraphTypes.GraphQLResponse = json.decodeFromString(raw.readText())
-            response.errors?.let {
-                printlnError("ERROR $it")
-                notificationsOperator.show(title = "Error from SourceGraph search result", body = "More info in application logs", type = ERROR)
-            }
-            response.data?.search?.results?.alert?.let {
+            raw.data?.search?.results?.alert?.let {
                 printlnError("ALERT $it")
                 notificationsOperator.show(title = "Alert from SourceGraph search result", body = "More info in application logs", type = WARNING)
             }
-            return response.data?.search?.results?.results.orEmpty().mapNotNull {
-                internalNameToSearchResult(searchHostName, it.getRepoName())
+            return raw.data?.search?.results?.results.orEmpty().mapNotNull {
+                val repoName = when (it) {
+                    is CommitSearchResult -> it.commit.repository.name
+                    is FileMatch -> it.repository.name
+                    is Repository2 -> it.name
+                }
+                internalNameToSearchResult(searchHostName, repoName)
             }.toSet()
         } catch (e: Exception) {
+            log.error("Exception caught running sourcegraph search", e)
             notificationsOperator.show(
                 title = "Search failed (More data in app logs)",
                 body = e.message ?: "Exception running search",
                 type = ERROR,
             )
-            e.printStackTrace()
             return emptySet()
         }
     }
 
-    private suspend fun rawSearch(searchHostName: String, settings: SourceGraphSettings, search: String): HttpResponse {
-        val client = httpClientProvider.getClient(searchHostName, settings)
-        return client.post("${settings.baseUrl}/.api/graphql?Search=") {
-            body = SourcegraphTypes.GraphQlRequest(SourcegraphTypes.SearchVaraibles(search))
-            timeout {
-                socketTimeoutMillis = 5 * 60 * 1000
-                requestTimeoutMillis = 5 * 60 * 1000
-            }
-        }
+    private suspend fun rawSearch(searchHostName: String, settings: SourceGraphSettings, searchString: String): GraphQLClientResponse<Search.Result> {
+        val client = getClient(searchHostName, settings)
+        val searchVars = Search.Variables(searchString)
+        val search = Search(searchVars)
+        return client.execute(search)
     }
+
+    private fun getClient(searchHost: String, settings: SourceGraphSettings) = GraphQLKtorClient(
+        url = URL("${settings.baseUrl}/.api/graphql"),
+        httpClient = httpClientProvider.getClient(searchHost, settings),
+
+    )
 
     private fun internalNameToSearchResult(searchHostName: String, name: String?): SearchResult? {
         val parts = name?.split('/')
@@ -70,8 +80,15 @@ class SourcegraphSearchClient(
     }
 
     suspend fun validateToken(searchHostName: String, settings: SourceGraphSettings): String = try {
-        val response: HttpResponse = rawSearch(searchHostName, settings, "count:1 timeout:10 f")
-        "${response.status.value}:${response.status.description}"
+        val response: GraphQLClientResponse<Result> = rawSearch(searchHostName, settings, "count:1 timeout:10s f")
+        when {
+            !response.errors.isNullOrEmpty() -> {
+                log.warn("Errors from SourceGraph GraphQL response {}", response.errors)
+                "ERRORS"
+            }
+            response.data?.search?.results?.results?.size == 1 -> "OK"
+            else -> "NO SEARCH RESULT"
+        }
     } catch (e: Exception) {
         e.printStackTrace()
         "Client error"